Moves evaluate to utils file
And makes the utils file a subset of the package, without directly
exporting it.
